requirements.txt
flet>=0.3.2
requests>=2.32.3
Pillow>=11.1.0
flet-cli>=0.10.0
uvicorn>=0.34.2
fastapi>=0.115.12
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
bcrypt>=4.0.1
psycopg2>=2.9.9
pydantic~=2.10.3